+++ changelog.md
@@ -1,2 +1,2 @@
 ### Connected Glass 1.1.14
-- Initial release of Connected Glass for Minecraft 1.21.9 & 1.21.11
+- Initial release of Connected Glass for Minecraft 1.21.11